refactor(web): move `relay.ts` to `relay/environment.ts`
